+++ config/dashbrd.php
@@ -6,7 +6,7 @@ use App\Enums\Period;
 
 return [
     'browsershot' => [
-        'timeout' => env('BROWSERSHOT_TIMEOUT', 60),
+        'timeout' => (int) env('BROWSERSHOT_TIMEOUT', 60),
         'node_binary' => env('BROWSERSHOT_NODE_BINARY', '/usr/local/bin/node'),
         'npm_binary' => env('BROWSERSHOT_NPM_BINARY', '/usr/local/bin/npm'),
     ],
